Update tests and implementations for index handling:
- Rename `shouldIgnoreMissingFieldsInDocument` to `shouldNotIgnoreMissingFieldsInDocument` and update test logic.
- Fix `IntegerRedisMessage.value()` to return `long` instead of `int`.
- Adjust `BucketCreateIndexHandlerTest` to include selector key in index creation template.
- Enhance `BucketDescribeIndexHandlerTest` for stricter index ID validation.
- Update `BucketListIndexesHandlerTest` to expect `primary-index` in the list of index names.

+++ kronotop/src/main/java/com/kronotop/server/resp3/IntegerRedisMessage.java
@@ -38,7 +38,6 @@ public final class IntegerRedisMessage extends AbstractNumberRedisMessage {
      * @return long value
      */
     public long value() {
-        return value.intValue();
+        return value.longValue();
     }
-
 }
